feat: modify the verifyAuth

+++ server/src/controllers/taskController.ts
@@ -2,6 +2,7 @@ import {Request, Response} from "express";
 // use asyncHandler to replace the try..catch..
 import asyncHandler from 'express-async-handler';
 import {TaskModel} from '../models/taskModel'
+import {UserModel} from "../models/userModel";
 
 /**
  * @desc get task
@@ -19,7 +20,9 @@ const getTask = asyncHandler(async (req: Request, res: Response) => {
 	}
 	*/
 
-	const tasks = await TaskModel.find();
+	// task is created by user so will find task by user
+	// match with TaskSchema user type
+	const tasks = await TaskModel.find({user: req.user._id});
 
 	res.status(200).json(tasks)
 })
@@ -32,7 +35,9 @@ const getTask = asyncHandler(async (req: Request, res: Response) => {
 const createTask = asyncHandler(async (req: Request, res: Response) => {
 
 	const task = await TaskModel.create({
-		text: req.body.text
+		text: req.body.text,
+		// the user create task
+		user: req.user._id
 	})
 	res.status(200).json(task)
 });
@@ -50,6 +55,23 @@ const updateTask = asyncHandler(async (req: Request, res: Response) => {
 		res.status(400);
 		throw new Error('Task not found')
 	}
+
+	const user = await UserModel.findById(req.user._id);
+
+
+
+	// check for user
+	if (!user) {
+		res.status(401)
+		throw new Error('User not found in this task')
+	}
+
+	// the task is not created by matched user
+	if (task.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('not matched user on this task')
+	}
+
 	const updatedTask = await TaskModel.findByIdAndUpdate(req.params.id, req.body, {new: true});
 
 	res.status(200).json(updatedTask)
@@ -67,6 +89,24 @@ const deleteTask = asyncHandler(async (req: Request, res: Response) => {
 	if (!task) {
 		res.status(400);
 		throw new Error('Task not found')
+	}
+
+
+	const user = await UserModel.findById(req.user._id);
+
+	console.log('user: ', user);
+
+	// check for user
+	if (!user) {
+		res.status(401)
+		throw new Error('User not found in this task')
+	}
+
+
+	// the task is not created by matched user
+	if (task.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('not matched user on this task')
 	}
 
 	await TaskModel.findByIdAndDelete(req.params.id);
